refactor: complete, emailAuth 리팩터링
