+++ src/features/auth/containers/SignupContainer.tsx
@@ -1,5 +1,4 @@
 import { useCallback } from 'react';
-
 import { useNavigate } from 'react-router-dom';
 
 import { Google } from '@/assets/images';
